Implemented Commandline Download Speedtest
Added the net/speedtest package that contains code for starting up a
speedtest server and a client. The speedtest command for starting the
server takes in a port number and maximum number of connections.
The speedtest command for starting a client takes in
parameters for the speedtest as well as the host and port of the
speedtest server to connect to. During a test, the client is measuring
download speeds, and the server is measuring upload speeds.

Signed-off-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>