derp: add optional rate limiting for DERP
Add rate limiting for DERP servers. This adds a naive hierarchical model for rate limiting by
having a server rate limiter, and a per client limiter. The per-client limiter is intended so
that a single client cannot hog all the server bandwidth, but is hopefully bursty enough to
allow for efficient use. The per-server limiter is intended to keep costs of running DERP down.

Signed-off-by: julianknodt <julianknodt@gmail.com>